BTreeMap: move generic functions out of navigate.rs
